fix: Ingredients, Ingredient, recipe num 타입 int로 수정

+++ sesac7_flutter_recipe_app/lib/data/mapper/ingredient_mapper.dart
@@ -14,7 +14,7 @@ extension IngredientDtoToModel on IngredientDto {
 extension IngredientModelToDto on Ingredient {
   IngredientDto toDto() {
     return IngredientDto(
-      id: id.toDouble(),
+      id: id.toInt(),
       name: name,
       image: image,
     );
